Updated file names to match new component name, updated so that the button is a regular anchor-tag with a href, adjusted Button component to be able to act as a regular link.
