+++ lib/Button/Button.js
@@ -68,7 +68,7 @@ function Button(props) {
   if (props.href) {
     return (
       // eslint-disable-next-line
-      <a className={getStyle()} onClick={handleAnchorClick} ref={getButtonRef} {...inputCustom}>
+      <a className={getStyle()} onClick={props.onClick && handleAnchorClick} ref={getButtonRef} {...inputCustom}>
         {children}
       </a>
     );
