+++ lib/structures/InfoPopover/InfoPopover.css
@@ -1,5 +1,5 @@
 /**
- * InformationIcon
+ * InfoPopover
  */
 
 @import "../../variables";
