+++ lib/structures/InfoPopover/stories/BasicUsage.js
@@ -3,7 +3,7 @@
  */
 
 import React from 'react';
-import InformationIcon from '../InformationIcon';
+import InformationIcon from '../InfoPopover';
 
 export default () => (
   <div style={{ padding: '15px' }}>
